+++ CHANGES.md
@@ -69,6 +69,7 @@ Changes in 12.0
 - Deprecate Portal Notes Files upload path $PortalNotesFilesPath global var in PortalPollsNotes.fnc.php
 - Use $FileUploadsPath . 'PortalNotes/' instead of $PortalNotesFilesPath in PortalNotes.php
 - Remove Portal Notes Attached Files Folder in assets/PortalNotesFiles/
+- Deprecate Assignments Files upload path $AssignmentsFilesPath global var in StudentAssignments.fnc.php
 
 Changes in 11.8.6
 -----------------